Feat: Added every fragment to the view package and made material3 working
