+++ app/src/main/java/nl/avans/drivioapp/adapter/MyCarsAdapter.kt
@@ -5,7 +5,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import nl.avans.drivioapp.MyCarsFragment
+import nl.avans.drivioapp.view.MyCarsFragment
 import nl.avans.drivioapp.databinding.ListElectricCarBinding
 import nl.avans.drivioapp.model.ElectricCar
