+++ app/src/main/java/nl/avans/drivioapp/view/EditBooking.kt
@@ -1,4 +1,4 @@
-package nl.avans.drivioapp
+package nl.avans.drivioapp.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,28 +9,29 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.setFragmentResultListener
 import androidx.fragment.app.viewModels
 import com.google.android.material.datepicker.MaterialDatePicker
-import nl.avans.drivioapp.databinding.FragmentEditAdvertisementBinding
+import nl.avans.drivioapp.R
+import nl.avans.drivioapp.databinding.FragmentEditBookingBinding
 import nl.avans.drivioapp.model.Advertisement
-import nl.avans.drivioapp.model.User
+import nl.avans.drivioapp.model.Reservation
 import nl.avans.drivioapp.model.User1
-import nl.avans.drivioapp.viewModel.AdvertisementViewModel
+import nl.avans.drivioapp.viewModel.ReservationViewModel
 import retrofit2.Response
 import java.text.SimpleDateFormat
 import java.util.*
 
-
-class EditAdvertisement : Fragment(R.layout.fragment_edit_advertisement) {
-    private var _binding: FragmentEditAdvertisementBinding? = null;
+class EditBooking : Fragment(R.layout.fragment_edit_booking) {
+    private var _binding: FragmentEditBookingBinding? = null;
     private val binding get() = _binding!!;
-    private val advertisementViewModel: AdvertisementViewModel by viewModels()
-    private lateinit var advertisement: Response<Advertisement>
-    private var advertisementId: Int = 0
+    private val reservationViewModel: ReservationViewModel by viewModels()
+    private lateinit var reservation: Response<Reservation>
+    private var reservationId: Int = 0
+
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentEditAdvertisementBinding.inflate(inflater, container, false);
+        _binding = FragmentEditBookingBinding.inflate(inflater, container, false);
         val view = binding.root;
         return view;
     }
@@ -38,26 +39,21 @@ class EditAdvertisement : Fragment(R.layout.fragment_edit_advertisement) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        setFragmentResultListener("advertisementIdEdit") { requestKey, bundle ->
-            advertisementId = bundle.getInt("advertisementIdEdit")
-            advertisementViewModel.getAdvertisementById(advertisementId)
+        setFragmentResultListener("reservationIdEdit") { requestKey, bundle ->
+            reservationId = bundle.getInt("reservationIdEdit")
+            reservationViewModel.getReservationById(reservationId)
         }
 
-        advertisementViewModel.getAdvertisementByIdResponse.observe(viewLifecycleOwner) {
-            advertisement = advertisementViewModel.getAdvertisementByIdResponse.value!!
+        reservationViewModel.getReservationByIdResponse.observe(viewLifecycleOwner) {
+            reservation = reservationViewModel.getReservationByIdResponse.value!!
 
-            val etTitle = binding.etTitle
-            val etDescription = binding.etDescription
-            val etPrice = binding.etPrice
+            val tvTitle = binding.tvTitle
             val etStartDate = binding.etStartDate
             val etEndDate = binding.etEndDate
 
-            etTitle.setText(advertisement.body()?.title.toString())
-            etDescription.setText(advertisement.body()?.description.toString())
-            etPrice.setText(advertisement.body()?.price.toString())
-            etStartDate.setText(advertisement.body()?.startDate.toString())
-            etEndDate.setText(advertisement.body()?.endDate.toString())
-
+            tvTitle.text = reservation.body()?.advertisement?.title.toString()
+            etStartDate.setText(reservation.body()?.startDate)
+            etEndDate.setText(reservation.body()?.endDate)
 
             val startDatePicker =
                 MaterialDatePicker.Builder.datePicker()
@@ -70,11 +66,21 @@ class EditAdvertisement : Fragment(R.layout.fragment_edit_advertisement) {
                     .build()
 
             etStartDate.setOnClickListener {
-                activity?.let { it1 -> startDatePicker.show(it1.supportFragmentManager, "DATE_PICKER") }
+                activity?.let { it1 ->
+                    startDatePicker.show(
+                        it1.supportFragmentManager,
+                        "DATE_PICKER"
+                    )
+                }
             }
 
             etEndDate.setOnClickListener {
-                activity?.let { it1 -> endDatePicker.show(it1.supportFragmentManager, "DATE_PICKER") }
+                activity?.let { it1 ->
+                    endDatePicker.show(
+                        it1.supportFragmentManager,
+                        "DATE_PICKER"
+                    )
+                }
             }
 
             startDatePicker.addOnPositiveButtonClickListener {
@@ -90,33 +96,28 @@ class EditAdvertisement : Fragment(R.layout.fragment_edit_advertisement) {
             }
 
             binding.btnConfirm.setOnClickListener {
-                val title = binding.etTitle.text.toString()
-                val description = binding.etDescription.text.toString()
-                val price = binding.etPrice.text.toString().toDouble()
                 val startDate = binding.etStartDate.text.toString()
                 val endDate = binding.etEndDate.text.toString()
 
-                val advertisement = Advertisement(
-                    advertisementId,
-                    title,
-                    description,
-                    price,
+                val reservationPut = Reservation(
+                    reservationId,
                     startDate,
                     endDate,
-                    User1(47)
+                    true,
+                    User1(47),
+                    Advertisement(reservation.body()?.advertisement?.advertisementId)
                 )
-                advertisementViewModel.putAdvertisementWithResponse(advertisement)
+                reservationViewModel.putReservationWithResponse(reservationPut)
             }
 
-            advertisementViewModel.putAdvertisementResponse.observe(viewLifecycleOwner) {
-                val response = advertisementViewModel.putAdvertisementResponse.value
+            reservationViewModel.putReservationResponse.observe(viewLifecycleOwner) {
+                val response = reservationViewModel.putReservationResponse.value
 
                 if (response?.code() == 200) {
                     Toast.makeText(activity, "Success!!", Toast.LENGTH_SHORT).show()
                 } else {
                     Toast.makeText(activity, "Failed!!", Toast.LENGTH_SHORT).show()
                 }
-
             }
         }
     }
