+++ app/src/main/java/nl/avans/drivioapp/view/StatisticDetailsFragment.kt
@@ -1,4 +1,4 @@
-package nl.avans.drivioapp
+package nl.avans.drivioapp.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,6 +8,7 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.setFragmentResultListener
 import androidx.fragment.app.viewModels
+import nl.avans.drivioapp.R
 import nl.avans.drivioapp.databinding.FragmentStatisticDetailsBinding
 import nl.avans.drivioapp.model.Statistic
 import nl.avans.drivioapp.viewModel.StatisticsViewModel
