+++ app/src/main/res/layout/fragment_advertisement_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".AdvertisementProfileFragment">
+    tools:context=".view.AdvertisementProfileFragment">
 
     <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
         android:layout_height="match_parent"
