+++ app/src/main/res/layout/fragment_edit_advertisement.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditAdvertisement">
+    tools:context=".view.EditAdvertisement">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
