+++ app/src/main/res/layout/fragment_edit_booking.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditBooking">
+    tools:context=".view.EditBooking">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
