+++ app/src/main/res/layout/fragment_my_car_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyCarDetailsFragment">
+    tools:context=".view.MyCarDetailsFragment">
 
     <TextView
         android:id="@+id/tvCarId"
