+++ app/src/main/res/layout/fragment_reservation_details.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ReservationDetailsFragment">
+    tools:context=".view.ReservationDetailsFragment">
 
     <TextView
         android:id="@+id/tvTitle"
